refactor(debug): isolate host dependencies
